eCAL C API module ecalc.cpp split in single units deprecated functions removed
