+++ ecal/core/include/ecal/cimpl/ecal_client_cimpl.h
@@ -122,18 +122,6 @@ extern "C"
   ECALC_API int eCAL_Client_AddResponseCallback(ECAL_HANDLE handle_, ResponseCallbackCT callback_, void* par_);
 
   /**
-    * @deprecated Please use eCAL_Client_AddResponseCallback instead
-    * @brief Add server response callback.
-    *
-    * @param handle_    Client handle.
-    * @param callback_  Callback function for server response.
-    * @param par_       User defined context that will be forwarded to the callback function.
-    *
-    * @return  None zero if succeeded.
-   **/
-  ECALC_API_DEPRECATED int eCAL_Client_AddResponseCallbackC(ECAL_HANDLE handle_, ResponseCallbackCT callback_, void* par_);
-
-  /**
    * @brief Remove server response callback. 
    *
    * @param handle_  Client handle.
